# sql: Improve efficiency of sorting queries in the presence of limiting