--- a/sql/testdata/order_by
+++ b/sql/testdata/order_by
@@ -55,6 +55,33 @@ SELECT a FROM t ORDER BY 1 DESC
 2
 1
 
+query ITT
+EXPLAIN SELECT a, b FROM t ORDER BY b LIMIT 2
+----
+0 limit count: 2, offset: 0
+1 sort  +b (top 2)
+2 scan  t@primary -
+
+query II
+SELECT a, b FROM t ORDER BY b DESC LIMIT 2
+----
+1 9
+2 8
+
+query ITT
+EXPLAIN SELECT DISTINCT a FROM t ORDER BY b LIMIT 2
+----
+0 limit    count: 2, offset: 0
+1 distinct
+2 sort     +b (iterative)
+3 scan     t@primary -
+
+query I
+SELECT DISTINCT a FROM t ORDER BY b DESC LIMIT 2
+----
+1
+2
+
 query II
 SELECT a AS foo, b FROM t ORDER BY foo DESC
 ----