We can improve the efficiency of in-memory sorting for `sortNodes` in both
the presence of a hard limit and the presence of a soft limit. This is accomplished
by adding two new sorting strategies to the `sortNode`, the first sorting only the
top-k when given a hard limit, and the second streaming the sort so that it can
terminate when a soft limit is reached at a higher `planNode`.

Given a query with `n` returned rows and a limit to `k` rows...

### Soft limit:

##### Example:
`SELECT DISTINCT v FROM kv ORDER BY v LIMIT 10`

##### Strategy:
Use a streaming heap sort, which can terminate once the soft limit is reached
by a wrapping `planNode`. This is achieved by using a min-heap to sort a
`valuesNode`. The heap is heapified initially, and for each call to `sortNode.Next`,
the next value is popped off the top of the heap and returned.

##### Time complexity:
The worst case time complexity of this approach remains `O(n log(n))`, but the average
time complexity of the sorting improves by a factor of `n/k`.

##### Space complexity:
The space complexity of the new approach will remain `O(n)`.

### Hard limit:

##### Example:
`SELECT v FROM kv ORDER BY v LIMIT 10`

##### Strategy:
Use a max-heap to keep track of the top `k` values as we iterate through the incoming
values from the wrapped `planNode`. We only need to make a copy of the values if the
new value is "less" than the value on the top of the heap. In this case, we can replace
the top value, keeping the max heap at size `k`. If a new value is not less than the
value at the top of the heap, we car drop it on the floor without making a deep copy and
without adding it to the `valuesNode`. At the end, we pop all values off the top of
the heap, which will sort the values in-place in the underlying slice.

##### Time complexity:
The worst case time complexity of this approach improves to `O(n log(k))`. This is
because all sorting operations work only on the size `k` heap.

##### Space complexity:
The space complexity of the new approach improves to `O(k)` because the `valuesNode` will
only ever grow to size `k`. The worst case number of values deep copies is still `n` times
if the incoming values are sorted in reverse and the heap is replaced each time. However,
on average the number of these deep copies will be reduced dramatically.


### Performance gains:

Note that below, `BenchmarkSort100000Limit10_Cockroach` will use a hard limit and
`BenchmarkSort100000Limit10Distinct_Cockroach` will use a soft limit.

```
name                                   old time/op    new time/op    delta
Sort100000Limit10_Cockroach-4             908ms ± 6%     815ms ± 7%  -10.31%  (p=0.000 n=8+8)
Sort100000Limit10Distinct_Cockroach-4     917ms ± 1%     810ms ± 6%  -11.67%  (p=0.001 n=6+8)

name                                   old alloc/op   new alloc/op   delta
Sort100000Limit10_Cockroach-4             143MB ± 0%     126MB ± 0%  -12.20%  (p=0.001 n=8+6)
Sort100000Limit10Distinct_Cockroach-4     143MB ± 0%     143MB ± 0%     ~     (p=0.081 n=8+6)

name                                   old allocs/op  new allocs/op  delta
Sort100000Limit10_Cockroach-4             1.81M ± 0%     1.75M ± 0%   -3.33%  (p=0.001 n=8+6)
Sort100000Limit10Distinct_Cockroach-4     1.81M ± 0%     1.81M ± 0%   -0.00%  (p=0.043 n=8+7)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5318)
<!-- Reviewable:end -->