@nvanbenschoten 2016-03-17
I made a number of pretty substantial changes here, including getting `explain (debug)` working by splitting the sorting types into their own `sortingStrategy` interface which allows iterative value population. This will allow us to play around with more efficient sorting algorithms in relative isolation later. PTAL.