@RaduBerinde 2016-03-17
This is awesome. The change looks great! Just some minor comments. We also need query tests for all strategies (you are probably working on that).

---

Review status: 0 of 3 files reviewed at latest revision, 5 unresolved discussions.

---

*[sql/sort.go, line 237 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4Jde5mQdY2EDirZk3:-KD4Jde5mQdY2EDirZk4:1584165349) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L237)):*
I am not sure this will always the case. Mabe if we somehow have a LIMIT inside a LIMIT we may want to have the outer limit trickle through and that would involve calling it twice (of course, still before the first Next). I would just remember the limit and initialize the strategy on the first Next (we're already doing a bit of initialization there).

---

*[sql/sort.go, line 347 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD3v-5Sc0XPnUnuMlnV:-KD3v-5TPdg3hIenzrr5:-983519564) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L347)):*
Add a comment explaining that through this interface we either retrieve values from a sorting strategy or from the planNode we wrap

---

*[sql/sort.go, line 362 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD3vN_gLDX82wkGn7mk:-KD3vN_gLDX82wkGn7ml:1298361681) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L362)):*
This usually comes first

---

*[sql/sort.go, line 405 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4KP41s5wxi_N-3Xw-:-KD4KP41s5wxi_N-3Xw0:-683792871) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L405)):*
I would just say "If the underlying node provides n rows and we produce k rows, the time complexity is O(n + k log n)".

---

*[sql/sort.go, line 471 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4L8FdcWJXsLZQsZXW:-KD4L8Fek0Ji_6_24N7M:-1448997782) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L471)):*
linear

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-:-KD4LXqqfsPL4cuUHohL:2024856499)*
<!-- Sent from Reviewable.io -->