@tamird 2016-03-17
Reviewed 1 of 1 files at r1, 1 of 3 files at r4, 2 of 2 files at r5.
Review status: all files reviewed at latest revision, 17 unresolved discussions.

---

*[sql/bench_test.go, line 674 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4MysFnEcFOohdKIvk:-KD4MysFnEcFOohdKIvl:-1002498122) ([raw file](https://github.com/cockroachdb/cockroach/blob/d9d712d35c92a60840b9044e9cb664598cf38367/sql/bench_test.go#L674)):*
something about the grammar here

---

*[sql/bench_test.go, line 697 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4N5iz2iQjFGIBynAU:-KD4N5iz2iQjFGIBynAV:381926458) ([raw file](https://github.com/cockroachdb/cockroach/blob/d9d712d35c92a60840b9044e9cb664598cf38367/sql/bench_test.go#L697)):*
you can hoist a lot of this out of the loop

also maybe just `var dist string` on this line

---

*[sql/sort.go, line 360 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4aEX6QhdmOOIrATBV:-KD4aEX6QhdmOOIrATBW:1169787613) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L360)):*
s/guarantees/guaranteed/ but also perhaps indicate that it should not be called more than once since this doc is for both implementers and consumers of this API.

---

*[sql/sort.go, line 366 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4aWjkUXdSMLvbTo7k:-KD4aWjkUXdSMLvbTo7l:733953765) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L366)):*
s/a quicksort/sort.Sort/

---

*[sql/sort.go, line 450 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4amYmIht6ruUOgzuU:-KD4amYmIht6ruUOgzuV:989614043) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/sort.go#L450)):*
nit: strconv.Itoa is more performant

---

*[sql/values.go, line 143 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4WjXFhKcic-1o_ozv:-KD4WjXFhKcic-1o_ozw:1414368453) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L143)):*
how come exported?

---

*[sql/values.go, line 170 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XJSlb0njidskJxqp:-KD4XJSlb0njidskJxqq:1943953574) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L170)):*
you should add a static assertion `var _ heap.Interface = (*valuesNode)(nil)`

---

*[sql/values.go, line 177 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XBggRQRSqskTcX7y:-KD4XBggRQRSqskTcX7z:1414368453) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L177)):*
how come exported?

---

*[sql/values.go, line 194 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XANfMlkrdrAKLoG6:-KD4XANfMlkrdrAKLoG7:1414368453) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L194)):*
how come exported?

---

*[sql/values.go, line 200 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XEIowKH5z8c6_2Z1:-KD4XEIowKH5z8c6_2Z2:-469765787) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L200)):*
ditto

---

*[sql/values.go, line 206 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XF53uTzy9DEEg6uX:-KD4XF53uTzy9DEEg6uY:-469765787) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L206)):*
ditto

---

*[sql/values.go, line 212 \[r5\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318#-KD4XFooBp-EdBP5oNMc:-KD4XFooBp-EdBP5oNMd:-469765787) ([raw file](https://github.com/cockroachdb/cockroach/blob/b4c509f6c648d5edb1c17c4ffb626b6176c91ffb/sql/values.go#L212)):*
ditto

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5318)*
<!-- Sent from Reviewable.io -->